SkylarkRepositoryContext: show rule location in case of wrong checksum
This location associated with the function call is the very position
of the call of the download method, i.e., it is a position somewhere
within implementation function of a repository rule. As such it does
not provide more information than the type of rule that was used.
Therefore, replace it by the location wher the rule was defined,
as this is more likely to be the position where the wrong hash came
from.

Closes #7353.

Change-Id: Ie3a2609377d926a619cc56bffc14aec26585d983
PiperOrigin-RevId: 234483022