Unify the options initialization and error handling in BuildEventServiceModule.
The error checking for the obtaining the options classes from the command environment
were scattered all over the class. Use a single method to initialize them and reset
them with every command.

Unifying these options uncovered a brittle dependency between BuildEventServiceModule and BazelBuildEventServiceModuleTest since it was relying on the streamer creation directly. The test now exercises the logic from BuildEventServiceModule.

PiperOrigin-RevId: 234489667