Ignore SIGCHLD in test setup script
SIGCHLD is expected when a subprocess ends, and is not indicative of a
failure. So, when trapping on signals to detect issues, do not trap on
this signal.

This has become important with Bash 5, as it has begun propagating this
signal up to this level.

Closes #7461.
Closes #7441.
Progress on #7410 when XML splitting is disabled.

PiperOrigin-RevId: 234568715