RepositoryResolvedEvent: represent modifications as named arguments
The RepositoryResolvedEvent was added to record the modified arguments a rule
returned in a resolved file. As such, they were thought of as dicts in Starlark
terms. However, the message this event produces is most useful for users who
wish to update their WORKSPACE file by hand. So make it more easy to copy and
paste.

Change-Id: I71e193a896c1d07cde3d5c1ebe1eb44eca52b659
PiperOrigin-RevId: 234591760