install-compile-source.md: hint at SOURCE_DATE_EPOCH
The `bazel` binary embeds a time stamp to indicate when it was
build. This is useful information in general, but makes the binary
not byte-for-byte reproducible. To overcome this problem, bazel
honors the environment variable SOURCE_DATE_EPOCH. Mention this
fact for the benefit of users packaging bazel, as reproducibility
is becoming a more and more important property of packages.

Change-Id: If983dfe977c5ed682d4efdb819c3a587dfbbe5c1
PiperOrigin-RevId: 234596785