Allow setting a default mtime in TarFileWriter so that all objects without explict mtime get the default. Context: Tests failing after https://github.com/bazelbuild/bazel/pull/7276
Next step: update rules_docker to use a default_mtime=0, so all tarballs are as they
were before.

Note: This should be put behind an incompatible change flag before the next bazel release.
Unfortunately, the cat is out of the bag in that we let the original PR changing the default_mtime went in alraeady.
PiperOrigin-RevId: 234803575