Change message about cycles in the WORKSPACE file
If a cycle is detected in the WORKSPACE file, i.e., a definition of a repository
is needed but not found till the place it is needed in the WORKSPACE file, this
can have two reasons,
- the definition could come later in the WORKSPACE file, or
- it could be missing completely.
Bazel has no chance of distinguishing those two cases, as the semantics of the
WORKSPACE file after this point depends on the not-found definition of the repository. So,
report both possibilities in the error message.

Closes #5697.

Change-Id: I92ff037c8b41c1fd78329db53800d11635b2e6e1
PiperOrigin-RevId: 234809390