Make Rule targets re-propagate AnalysisFailureInfo automatically
Under --allow_analysis_failures, if a rule has any dependencies that propagate AnalysisFailureInfo, then the rule itself is almost certain to fail, and thus will re-propagate AnalysisFailureInfo automatically containing the specific failures of its dependencies, instead of an AnalysisFailureInfo that is due to some more-obscure exception.

Progress toward #6237. This implements a planned portion of the design that was previously forgotten.

This is technically a breaking change, but should be very low-risk. (Previously, it would be possible to depend on a failing target, not use any of its data, and therefore not fail. After this change, this is no longer possible.)

RELNOTES: None.
PiperOrigin-RevId: 234864320