Use exec path as a fallback when sorting by root-relative path.
ImmutableSortedMap uses the comparator it's given to check for uniqueness;
in android_local_test scenarios where the same manifest is picked up both
through the normal configuration and the Android configuration, this can
result in duplicates.

Instead of considering the two artifacts equal in this case, fall back to
comparing their configuration directories.

RELNOTES: None.
PiperOrigin-RevId: 234879527