Automated rollback of commit ab72246c4498a80a703771e267f5c1ebbc3fbf4b.
*** Reason for rollback ***

Partial rollback because we need to migrate some folks before we can get the new functionality live.

*** Original change description ***

Support Xcode betas with xcode-locator

Previously if you had 2 builds of the same version of Xcode, for example
10.2.0 beta 1 and 10.2.0 beta 2, there was no way to disambiguate them.
This change appends the `ProductBuildVersion` of each Xcode version to
allow you to specify that if desired.

Closes #7371.

PiperOrigin-RevId: 234900481