Simple fix for #7083.
This caches the results of `#createOutputArtifact()` inside RuleContext, in a provider. `LocationExpander#buildLocationMap()` then uses this provider/cache instead of calling `#createOutputArtifact()` manually.

There's likely a better way of doing this, but for now this unblocks us in our time of need. I'll keep digging to try & locate a better approach.