new_local_repo: fix correctness bug with glob()
In Skyframe, the DirectoryListingStateValue (DLS)
of an external repository's path will now depend
on the RepositoryDirectoryValue (RD).

Example: the DLS of [output_root]/external/X now
depends on the RD of `@X`.

This is relevant for `glob()` expressions in the
top-level package of the repository. The GlobValue
depends on the DLS in order to get notified about
added/removed files. In the root of external
directories, those files are added/removed by the
repository function. Hence the DLS should depend
on the evaluation of the repository function, i.e.
the creation of the RD. And that is exactly what
this commit now does -- it ensures this dependency
is declared.

The bugfix is specific to Windows. On Linux,
GlobValue depends on the DLS, but it also depends
on symlinks in the directory (to get notified
about a directory symlink getting replaced by a
file with the same name, which would affect
recursive globbing). And the symlinks depend on
the RD, so the GlobValue always depended on the
RD.

However on Windows the files in the root of the
repository are copies, not symlinks, and the Glob
does not depend on files (as a means of
optimization), so on Windows the GlobValue didn't
depend on the RD, which caused issue #6351.

More info: https://github.com/bazelbuild/bazel/issues/6351#issuecomment-465488344

Fixes https://github.com/bazelbuild/bazel/issues/6351

Closes #7473.

PiperOrigin-RevId: 234938677