Clean up args conversion to an argv array.
This introduces a new CharPP class that implements the conversion of
C++ data structures to their corresponding char** representation
needed for execv-like calls.

As a side-effect of this change, we fix two issues: first, a memory
leak; and second, a theoretical const-correctness violation.

I am doing this change to provide room for implementing a similar
conversion function for envp arrays, which I will need to switch to
using posix_spawn.

Prerequisite cleanup to address
https://github.com/bazelbuild/bazel/issues/7446.  While this does
fix an issue with a memory leak, this change should have no visible
effects.

RELNOTES: None.
PiperOrigin-RevId: 234968622