Move OutputStream creation out of FileTransport and into the BuildEventServiceModule.
A FileTransports should be created iff the file provided by the user
using --build_event_{text,binary,json}_file is actually writable.

In turn, there isn't really a reason why not being able to write to the path provided should fail the build; however, changing the error to a warning would be an incompatible change. So punting it for now.

As a side effect, slightly improve the unit tests for the transport classes by using the setUp and tearDown methods properly.

PiperOrigin-RevId: 235020625