Split ActionStatusMessage into separate events.
ActionStatusMessage had a very strange API in the sense that it carried
multiple different statuses but these were not explicitly represented
anywhere: callers had to guess what status was represented by a combination
of the values of the fields in the message.

To fix this, split the status updates into separate events with good names
and types.  This keeps each event restricted to the data it has to carry,
preventing callers from having to guess what is or is not available, and
matches what we do for many of the other action-related events.

Note that it'd be slightly nicer to use an enum of some form so that callers
could be guaranteed to handle all possible status cases... but doing so
results in even more convoluted code because each status needs to carry
different information around.

This change highlights two more deficiencies that I am fixing.  First,
there is no such thing as a "preparing" event separate from an "started"
event so I removed the former representation.  Second, the "messages" in
the status messages were used to update the UI, thus conflating internal
status representations with the specific ways in which they should be
displayed to the users; I just pushed the formatting responsibility to
the UI modules instead.

Necessary cleanup to address https://github.com/bazelbuild/bazel/issues/7345.

RELNOTES: None.
PiperOrigin-RevId: 235051214