Fix #7515: bazel_rules_test fails on Windows.
The test had a couple of problems:
- A misunderstanding of how PATH looks like in MSYS on Windows - it's not delimited by ';', but by ':'. Drive letters are not a problem, because MSYS represents these as virtual directories (e.g. C:\TEST becomes /c/TEST). By using a ';', the first entry in the PATH became invalid, which lead to the Bazel wrapper no longer being found.
- "export env variables and then try to restore them later" instead of just calling exactly the program that should use them with the env vars explicitly set. This is problematic, because the env vars do not get restored in case of failure and also because it's overly broad.
- Rewriting the existing tool "env" using "echo" and a select() construct.
- Didn't redirect output of "bazel" into $TEST_log.

By setting PATH to a wrong value on Windows, this caused the "bazel" command in the test to accidentally run the "bazel" binary from the system instead of the runfiles, so the test didn't test the correct Bazel version. We only noticed this when we replaced the "bazel" binary on the CI machines with Bazelisk, which crashed, because $HOME (Linux, macOS) and %LocalAppDir% were not set and Bazelisk requires these.

RELNOTES: None.
PiperOrigin-RevId: 235490629