Mitigate unordered dirtying of rewound nodes by allowing repeats
When a Skyframe node is restarted, and it dirties some of its transitive
dependencies, it does not dirty those dependencies as a single atomic
action, nor does it dirty them in a reverse topological order (i.e.
starting from the leaves of the dirtied subgraph and proceeding to the
subgraph's root(s)).

This leaves rewinding vulnerable to a race where e.g. a sibling action
requests the to-be-dirtied subgraph's root artifact node (corresponding
to the lost input) after it's been dirtied, but before its child
generating action node has been dirtied.

That artifact node may then evaluate to completion, and become done,
before its child is dirtied. Then, when the failed action node (which
requested the rewinding) restarts, and requests its inputs, it will find
the lost input's node is done and proceed with its evaluation, despite
that its lost input's generating action has yet to run.

A proper fix will ensure that no node in a rewinding's to-be-dirtied
subgraph is scheduled for evaluation before its transitive deps in that
subgraph are dirtied.

Drive-by limiting of action toString length in rewinding log statements.

RELNOTES: None.
PiperOrigin-RevId: 235625329