Remove toolchain resolution logic from find_java_toolchain helper
and put it exclusively in the implementation of the toolchain alias rules.
This is similar to the removal of the toolchain resolution logic from the
native rules in https://github.com/bazelbuild/bazel/commit/24bbdb518d70143eb8575f9e9cb61545e935d8cf. The indirection through the aliases is
necessary to force a host configuration transition for the host runtime
alias.

PiperOrigin-RevId: 235721677