Refactor the async execution interface for actions
This makes the interface generic such that it can be used for multiple
steps as well as for multiple concurrently running spawns, effectively
allowing any action subclass to opt-in to the new framework.

The new API is directly equivalent to the old execute method, so we can
subsequently switch to using the new API for all actions, even if
they're still executing synchronous, and then move individual action
types over piecemeal.

Progress on #6394.

PiperOrigin-RevId: 235726176