Add a python_version attribute to py_runtime and its associated provider
This is used to declare whether a runtime is for Python 2 or Python 3. When we move to toolchains, the runtime's declared version will be validated against the py_binary's version. For compatibility, validation will not occur for the old --python_top way of specifying runtimes.

Initially the attribute is optional on the py_runtime rule and mandatory on the PyRuntimeInfo provider. At some point we'll add an incompatible change to make the attribute mandatory on the rule too. In the meantime it defaults to a version in the same way as py_binary, i.e. with --incompatible_py3_is_default.

Work toward #7375.

RELNOTES: `py_runtime` gains a `python_version` attribute for specifying whether it represents a Python 2 or 3 interpreter.
PiperOrigin-RevId: 235774497