Remove ACTION_EXECUTE profiling
This task heavily overlaps the existing ACTION task to the extent
that both are basically equivalent, and ACTION_EXECUTE was already
excluded by the Json profiler config.

While I'm here, also make sure that errors for async actions are
reported correctly. We'll get test coverage as we switch actions over
to the new API.

PiperOrigin-RevId: 235884902