Refactor StandaloneTestStrategy with new interfaces
This change adds a set of generic interfaces to TestStrategy, in terms
of which all existing strategies can be implemented. It also refactors
StandaloneTestStrategy to do so.

This increases sharing at the test strategy level, which is in
preparation for supporting async test execution.

Note that this fixes the collection of SpawnResult instances as a side
effect - previously we only kept the last SpawnResult, now we collect
all of them.

PiperOrigin-RevId: 235914420