Use a reasonable set of spawn strategies by default.
This is only enabled when using --incompatible_list_based_execution_strategy_selection.

Note that this does *not* yet disable the built-in fallback behavior in the spawn strategies, but that's not required for this feature to work. That'll just be a nice addition to make the code clearer and allows us to remove the fallback code from all strategies.

RELNOTES: When using --incompatible_list_based_execution_strategy_selection, Bazel will use remote execution by default (if you specify --remote_executor), otherwise persistent workers (if the action supports it), otherwise sandboxed local execution (if the action and platform supports it) and at last unsandboxed local execution. The flags --spawn_strategy and --strategy continue to work as before - this only sets new defaults for the case where you don't specify these flags.
PiperOrigin-RevId: 235932751