Dirty rewinding subgraphs in reverse topological order
The ActionRewindStrategy is aware of the graph structure for the nodes it
asks to dirty. By conveying that graph structure to Skyframe's rewinding
engine, the strategy ensures that artifacts are dirtied after their
generating action is dirtied, and that the transitive dependencies of
propagating actions and Skyframe-aware actions are dirtied before their
parent action is dirtied.

Skyframe's rewinding engine dirties nodes ordered via a postorder DFS
traversal of that graph structure. This prevents a parent node from
getting scheduled, consuming the value of a done child node, and
completing, before the child is dirtied, when the rewinding engine
is dirtying both the parent and child.

SkyframeAwareActions now supply the rewinding strategy with a graph of
the nodes they need rewinding to dirty, instead of just a set. The graph
may include edges to artifacts, whose generating actions must be checked
for Skyframe-awareness and propagation, in the same way as the
generating actions of input artifacts of propagaing actions must be
checked. This CL refactors this BFS traversal in ActionRewindStrategy to
be more symmetrical with respect to those two sources of artifacts.

This CL includes a drive-by fix to emitting invalidations via
EvaluationProgressReceiver.invalidated during rewinding.

RELNOTES: None.
PiperOrigin-RevId: 235955000