Add unit tests that check what interpreter py_binary invokes
This adds analysis-time test cases that assert on the Python interpreter path that gets emitted to the stub script of a py_binary or py_test target. This will be useful as we start modifying this behavior using Python toolchains (#7375).

Also renamed PyRuntimeTest for consistency with other "ConfiguredTargetTest"s.

Work toward #7375.

RELNOTES: None
PiperOrigin-RevId: 235970443