Fix building bazel without Xcode
When changing run_xcode_locator to not return an error and call `fail()`
instead we didn't realize that it was called outside of
xcode_configure.bzl, where it Xcode not being installed would already be
handled by a call to `xcodebuild -version` below. Since it's also called
directly from osx_cc_configure.bzl, we need to allow that case to not
fail the configuration, but also not setup Xcode. In this case returning
`None` will fall back to the generic unix setup.