Batch the recursive GlobValue dep with the DirectoryListingValue dep.
This is a strict win wrt parallelism, at the cost of code ugliness. Now it takes less total Skyframe overhead to get to the recursive per-dirent work. Previously, we had to sequentially wait for a Skyframe thread to be available for the recursive GlobValue dep, and then a Skyframe thread to be available for the GlobValue restart, and then a Skyframe thread to be available for the DirectoryListingValue work.

RELNOTES: None
PiperOrigin-RevId: 236182538