Replace ExitFunction with BuildEventServiceAbruptExitCallback to properly report transport failures back to the BuildEventServiceModule.
The purpose of deleting this function is twofold:
i) Unify the way we report errors to the main thread since now all the transports
use the same callback.
ii) Separate UI messages from actions that fail the build (ModuleEnvironment.exit).

With these improvements we're one step closer of making the BES module the
sole responsible of posting messages to the CommandEnvironment.EventHandler
(instead of letting the transports or the streamer post messages too which has
caused deadlocks in the past). Additionally, we greatly simplify the handling
of --bes_results_url and improve our UI by printing more clear messages.

PiperOrigin-RevId: 236386523