Add SpawnContinuation
Async action execution is implemented as layered, nested state machines,
where each state machine corresponds to a specific behavior. Each
instance of a continuation is a state in a state machine, and some
states run entire state machines at the next level.

There are now 4 layers of continuations:

1. The ActionExecutionFunction uses a continuation for each input-discovering
action to reduce the overhead of function restarts.

2. The SkyframeActionExecutor uses a continuation for each group of
shared actions (as determined by the name of the primary output). Note
in particular how this differs in scope from the first layer.

3. Action provides a generic level of continuations for actions to use
to represent any sequence of internal work needed for that action.
SpawnAction specifically uses this to represent the additional steps
wrapping the lower-level spawn execution.

4. The SpawnContinuation represents the execution of a single spawn.
This includes fallback from remote to local execution, as well as the
work necessary to implement tree artifacts. Tree artifacts may be
implemented by wrapping the given spawn in a shell script that unpacks
any tree artifacts present in its inputs, as well as a subsequent spawn
that unpacks any tree artifacts present in its outputs (post spawn
execution).

We could consider merging some of these layers. However, it is not
trivial to extend or modify an existing state machine.

For example, it is not straightforward to merge 3 and 4. Layer 4
represents the state machine required by the SpawnActionContext, which
may differ based on the locality (remote or local) or sandboxing
behavior of that strategy, whereas 3 represents the state machine
required by the action. Merging them would require that actions are
aware of all possible state machines required for any
SpawnActionContext, or alternatively pass some information through to
the SpawnActionContext (a continuation?) that is then transitioned to
after the strategy-specific state machine is completed.

While this may save us some memory, it may also end up increasing
complexity.

Progress on #6394.

PiperOrigin-RevId: 236597242