Respect topological order for pattern expansion failure events.
When invoking Bazel with -k and successful and failing patterns are
combined, a top-level PatternExpanded event is generated with one or
more PatternExpandedErrors underneath. The latter are also attached
through Progress nodes.

It has been observed that topological order isn't fully respected for
these events. Bazel may first emit the Progress nodes, followed by the
PatternExpandedErrors, only to be followed by the main PatternExpanded
node. This causes errors/warnings when displaying these events through
Buildbarn Browser, as the parser there requires a topological order to
ensure the resulting graph is acyclic.

Change PatternExpandingError to implement BuildEventWithOrderConstraint.
Provide a postedAfter() that requires PatternExpandedErrors to be
preceded by the main PatternExpanded.

Related to: #7608