Slightly clarify SandboxfsSandboxedSpawn code.
The two helper functions reconfigure() and createMappings() were intended
to be static from the beginning, but some reason weren't.  This was
confusing because these functions took an "inputs" parameter... but this
parameter was also available as a class attribute.

Fix this by folding the contents of reconfigure() into its only caller,
createFileSystem(), and then make createMappings() static.

RELNOTES: None.
PiperOrigin-RevId: 237006184