Remove BuildOptions.NULL_OPTIONS.
This was conceived as a way to make the null configuration look very same to any other configuration in that it's the result of a BuildConfigurationFunction, however, this would have resulted in an extra outgoing Skyframe edge from every InputConfiguredTarget, which is a bit too much overhead for some code beauty.

I also tried to make it so that NullTransition.INSTANCE is special-cased in ConfigurationResolver#resolveConfigurations() and thereby eliminate special-casing the null configuration in DependencyResolver#fullyResolveDependencies(). That one was sunk by the fact that #resolveConfigurations() is not called when the configuration of the configured target being analyzed is null, which means a little more garbage.  It's probably doable, but let's do it in a followup change.

RELNOTES: None.
PiperOrigin-RevId: 237008737