Implement combined disk and HTTP cache
This PR adds the ability to enable disk and HTTP cache simultaneously. Specifically, if you set
```
build --remote_http_cache=http://some.remote.cache
build --disk_cache=/some/disk/cache
```
Then Bazel will look for cached items in the disk cache first. If an item is not found in the disk cache, it will be looked up in the HTTP cache. If it is found there, it will be copied into the disk cache. On put, Bazel will store items in both the disk and the HTTP cache.

We tested this change on a relatively large internal project and found that a fully cached, clean build was 5-10 times faster with the mixed cache (i.e. reading from disk cache) than with a pure remote cache, depending on the remote cache location.

cc @dajmaki @francesco-da

Closes #7512.

PiperOrigin-RevId: 237011131