Refactor "determine all dependency edges from all aspects" logic to be a streaming visitation.
In doing so, we use more efficient algorithm that notably doesn't have several large sources of garbage. The inefficiencies and garbage churn of the old algorithm are more easily noticed when you consider the pseudo-code.

The old algorithm for was:

  res <- []
  for each attribute A of T that entails a label dep
    for each label dep L entailed by A
      aspectDeps <- {}
      for each aspect S of A
        if S is satisfied by T
          for each attribute A' of S
            for each label dep L' entailed by A'
              // Note that L' has nothing directly to do with L.
              aspectDeps <- aspectDeps U {A', L')
      for each {a, l} in aspectDeps
        res <- res + [l]
  return res

And the new algorithm is:

  res <- []
  for each attribute A of T that entails a label dep
    for each aspect S of A
      if there exists a label dep L entailed by A that satisfies S
        for each label dep L' entailed by S
          res <- res + [L']
  return res

This refactor exposes an existing inefficiency in LabelVisitor. I added a TODO comment explaining it.

RELNOTES: None
PiperOrigin-RevId: 237063485